Fix to not require the `village.json` schema to determine version numbers.
Due to using the JSON in the configuration phase, it meant that users of the
library can't build the library without the file being present. This is not
ideal. So the version of the spec that the Open API implementations supports
is not a project property and will need to be manually updated.